Fixed some thread safety issues, implemented report saving and listing created reports.

+++ backend/SkripsiAppBackend/SkripsiAppBackend/Persistence/Models/Report.cs
@@ -10,6 +10,6 @@
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
 
-        public int Expenditure { get; set; }
+        public long Expenditure { get; set; }
     }
 }
